Use common function showing datasource info
